distill

+++ lift/gated_memory/model.py
@@ -30,7 +30,7 @@ class MyLayerNorm(nn.Module):
 
     def forward(self, hidden_states):
         input_dtype = hidden_states.dtype
-        hidden_states = hidden_states.to(torch.float32)
+        # hidden_states = hidden_states.to(torch.float32)
         hidden_states = hidden_states - hidden_states.mean(-1, keepdim=True)
         variance = hidden_states.pow(2).mean(-1, keepdim=True)
         hidden_states = hidden_states * torch.rsqrt(variance + self.variance_epsilon)
